+++ src/nextjs/index.ts
@@ -1,0 +1,2 @@
+export { default as getLanguageStaticProps } from "./getLanguageStaticProps";
+export { default as getMenuStaticProps } from "./getMenuStaticProps";