+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - PageContent component
 - Navigation component
 - Navigation component that can request its own data in an Apollo project
+- NextJS getLanguageStaticProps and getMenuStaticProps utilities that can be used to fetch hydrated data when using static page generation
 
 ### Fixed